ENT-1880: Port OwnerProductResource to spec-first
