+++ server/src/test/java/org/candlepin/resource/ProductResourceTest.java
@@ -210,7 +210,7 @@ public class ProductResourceTest extends DatabaseTestFixture {
         this.setupDBForOwnerProdTests();
 
         List<AsyncJobStatusDTO> jobs = productResource
-            .refreshPoolsForProduct(Collections.singletonList("p1"), true)
+            .refreshPoolsForProducts(Collections.singletonList("p1"), true)
             .collect(Collectors.toList());
 
         assertNotNull(jobs);
@@ -228,7 +228,7 @@ public class ProductResourceTest extends DatabaseTestFixture {
         this.setupDBForOwnerProdTests();
 
         List<AsyncJobStatusDTO> jobs = productResource
-            .refreshPoolsForProduct(Arrays.asList("p1", "p2"), false)
+            .refreshPoolsForProducts(Arrays.asList("p1", "p2"), false)
             .collect(Collectors.toList());
 
         assertNotNull(jobs);
@@ -246,7 +246,7 @@ public class ProductResourceTest extends DatabaseTestFixture {
         this.setupDBForOwnerProdTests();
 
         List<AsyncJobStatusDTO> jobs = productResource
-            .refreshPoolsForProduct(Collections.singletonList("p3"), false)
+            .refreshPoolsForProducts(Collections.singletonList("p3"), false)
             .collect(Collectors.toList());
 
         assertNotNull(jobs);
@@ -264,7 +264,7 @@ public class ProductResourceTest extends DatabaseTestFixture {
         this.setupDBForOwnerProdTests();
 
         List<AsyncJobStatusDTO> jobs = productResource
-            .refreshPoolsForProduct(Collections.singletonList("nope"), false)
+            .refreshPoolsForProducts(Collections.singletonList("nope"), false)
             .collect(Collectors.toList());
 
         assertNotNull(jobs);
@@ -274,7 +274,7 @@ public class ProductResourceTest extends DatabaseTestFixture {
     @Test
     public void testRefreshPoolsByProductInputValidation() {
         assertThrows(BadRequestException.class, () ->
-            productResource.refreshPoolsForProduct(new LinkedList<>(), true)
+            productResource.refreshPoolsForProducts(new LinkedList<>(), true)
         );
     }
 
@@ -299,7 +299,7 @@ public class ProductResourceTest extends DatabaseTestFixture {
 
         List<AsyncJobStatusDTO> jobs = null;
         try {
-            jobs = productResource.refreshPoolsForProduct(Collections.singletonList("p1"), true)
+            jobs = productResource.refreshPoolsForProducts(Collections.singletonList("p1"), true)
                     .collect(Collectors.toList());
         }
         catch (Exception e) {
